Changed name of the windows snapshots.
The name of the generated *.zip and *.exe does not depend on the date
of build. This change fixes some non trivial crashes (poping up when
the date change but not the commit hash or vice versa). Note that the
unarchived zip or exe installer will be in a directory named with the
build date.